Move global variables to NifflerRuntime
